services/gateway-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
httpx==0.27.2
python-jose[cryptography]==3.3.0